refactor: prefer instrumentation-client
